Move DailyTaskStartDto into common/dto, derive net table entry from it
