+++ src/common/dto/daily-task.d.ts
@@ -42,3 +42,19 @@ export interface DailyTaskResultDto {
   /** 本局归属的任务日，取自 /game/start 响应，用于服务端判断任务归属日期 */
   dayId: string;
 }
+
+/** /game/start 响应里 dailyTasks 数组的单项形状（每个 steamId 一条） */
+export interface DailyTaskStartDto {
+  steamId: number;
+  dayId: string;
+  candidates: TaskCandidateDto[];
+  completedTasks: CompletedTaskDto[];
+  todaySeasonPoint: number;
+  history: DailyTaskHistoryEntryDto[];
+}
+
+/** daily_task net table 单个玩家条目 */
+export interface DailyTaskNetTableEntry extends Omit<DailyTaskStartDto, 'steamId' | 'dayId'> {
+  enabled: boolean;
+  selectedTaskId?: string;
+}
